fix: allow a fresh install before a domain certificate exists
A fresh Direct/external-proxy install failed because the generated
sing-box server config and the install-time certificate pin both
referenced a certificate that is only obtained after install. Default
the protocol-listener certificate mode to self-signed (generated at
install, clients skip verification) and make the Direct install pin
tolerant of a not-yet-obtained certificate.

+++ src/qr.rs
@@ -9,7 +9,7 @@ use qrcode::{Color, QrCode};
 pub fn render_ansi(data: &str) -> Result<String, String> {
     let code = QrCode::new(data.as_bytes()).map_err(|error| error.to_string())?;
     let modules = code.to_colors();
-    let width = code.width() as usize;
+    let width = code.width();
     let mut out = String::new();
     out.push_str("\x1b[30;47m");
     let mut row = 0;
